requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
tensorflow
numpy
pillow
h5py
